refactor: split automation session module
